+++ src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/DynamicClient/DynamicStreamingHubClientBuilder.cs
@@ -1,6 +1,7 @@
 #if NON_UNITY || ((!ENABLE_IL2CPP || UNITY_EDITOR) && !NET_STANDARD_2_0)
 
 using Grpc.Core;
+using MagicOnion.Internal.Reflection;
 using MagicOnion.Serialization;
 using MagicOnion.Server.Hubs;
 using MagicOnion.Utils;
