+++ src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/MagicOnion.Shared/Internal/Buffers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d3d69d225c61e4c4a9c8d5c924ea6341
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
