+++ src/MagicOnion.Client.Unity/Assets/Scripts/MagicOnion/MagicOnion.Client/MagicOnion.Shared/Internal/Reflection.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6dc40ea58fc3d634da4c4f98218069d0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
