+++ src/MagicOnion.Client/DynamicClient/DynamicClientAssemblyHolder.cs
@@ -1,5 +1,5 @@
 #if NON_UNITY || ((!ENABLE_IL2CPP || UNITY_EDITOR) && !NET_STANDARD_2_0)
-using MagicOnion.Utils;
+using MagicOnion.Internal.Reflection;
 
 namespace MagicOnion.Client.DynamicClient
 {
