+++ src/MagicOnion.Server/Hubs/Group.ConcurrentDictionary.cs
@@ -1,10 +1,10 @@
 using MagicOnion.Serialization;
 using MagicOnion.Server.Diagnostics;
-using MagicOnion.Utils;
 using MessagePack;
 using System.Collections.Concurrent;
 using System.Diagnostics.CodeAnalysis;
 using Microsoft.Extensions.Logging;
+using MagicOnion.Internal.Buffers;
 
 namespace MagicOnion.Server.Hubs;
